# archive/v3_cleanup_backup/cbs-platform/platform/api-schema/openapi-v2.yml
openapi: 3.0.3
info:
  title: CBS_PYTHON V2.0 - Core Banking System API
  description: |
    Advanced microservices-based Core Banking System with comprehensive financial services.
    
    ## Features
    - Microservices architecture with domain-driven design
    - OAuth2 and API key authentication
    - Real-time event streaming
    - Advanced rate limiting and throttling
    - GraphQL support for complex queries
    - Webhook notifications
    - Comprehensive audit logging
    - Multi-tenancy support
    
    ## Security
    - JWT-based authentication
    - Role-based access control (RBAC)
    - API key management
    - Request/response encryption
    - PCI DSS compliance
    
  version: 2.0.0
  contact:
    name: CBS Engineering Team
    url: https://cbs-platform.com/support
    email: api-support@cbs-platform.com
  license:
    name: Proprietary
    url: https://cbs-platform.com/license
servers:
  - url: https://api.cbs-platform.com/v2
    description: Production API
  - url: https://staging-api.cbs-platform.com/v2
    description: Staging API
  - url: http://localhost:8000/v2
    description: Development API
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  - OAuth2: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT token
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh JWT token
      description: Refresh expired JWT token
      operationId: refreshToken
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate JWT token
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # Customer Management
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
    
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.cbs-platform.com/oauth/authorize
          tokenUrl: https://auth.cbs-platform.com/oauth/token
          scopes:
            read: Read access to resources
            write: Write access to resources
            admin: Administrative access
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
        - request_id
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        request_id:
          type: string
          format: uuid
          description: Unique request identifier
...
    SuccessResponse:
      type: object
      required:
        - success
        - message
        - timestamp
        - request_id
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          description: Success message
        data:
          type: object
          description: Response data
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Username or email
        password:
          type: string
          format: password
          description: User password
        remember_me:
          type: boolean
          default: false
          description: Extended session duration
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: Refresh token
    AuthResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: Refresh token
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Token expiration time in seconds
        scope:
          type: string
          description: Token scope
        user:
          $ref: '#/components/schemas/UserInfo'
    UserInfo:
      type: object
      required:
        - id
        - username
        - roles
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string

    # Customer Schemas
...